Splitting handlers into different files.
I have also created a top-level utils.py for notebook related stuff
